third_party/WebKit/LayoutTests/http/tests/websocket/handshake-fail-by-more-accept-header-expected.txt

CONSOLE ERROR: line 21: WebSocket connection to 'ws://localhost:8880/handshake-fail-by-more-accept-header' failed: Error during WebSocket handshake: 'Sec-WebSocket-Accept' header must not appear more than once in a response
Test that WebSocket handshake fails if there are more one Sec-WebSocket-Accept header field in the response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS successfullyParsed is true

TEST COMPLETE
